refactor(ts-storage-duckdb): repartition shared mod tests into per-entity test mods
Moves the 44-test shared `mod tests` block out of lib.rs into each
entity's `#[cfg(test)] mod tests`, finishing the entity-based layout
the previous commit committed to.

Classification (by primary tested method):
  exchanges.rs      3   http_exchange_*
  calls.rs          7   test_write_calls_*, test_query_calls_*, test_query_call_by_id
  schema.rs         2   test_init_*
  metrics.rs       19   test_write_metrics_*, query_finish_reasons_*,
                        test_query_metrics_*, test_multi_source_*,
                        query_distinct_finish_reasons_returns_pairs
  distincts.rs      3   test_query_distinct_*
  util.rs          10   test_build_dimension_where_* (pure unit, no DB)

Fixtures (in_memory, sample_call, sample_metric) are duplicated per
entity rather than exposed cross-mod — same precedent as the existing
turns.rs::tests and retention.rs::tests, which have their own
sample_turn / mk_call helpers. Sharing would have required
`pub(crate) mod tests { pub(crate) fn ... }` ceremony that obscures
the test boundary; per-entity duplication is a few literal-data
builders.

lib.rs: 1875 → 232 lines. Now only DuckDbBackend struct + open() +
delegating impl StorageBackend + mod declarations.

No behavior change; 64 tests pass (same total as before the move).
